initial commit, just finished dialog

+++ ContentBuilder/ContentProcessor/CramTextureAtlasData.cs
@@ -44,7 +44,7 @@ namespace ContentProcessor
 	{
 		public string[] Frames { get; set; }
 
-		public int FrameRate { get; set; }
+		public float FrameRate { get; set; }
 		public int XOrigin { get; set; }
 		public int YOrigin { get; set; }
 	}
